+++ src/remix.cpp
@@ -46,7 +46,7 @@ ReMix::ReMix(QWidget* parent) :
 
     Logger::setInstance( new Logger( this ) );
     if ( Settings::getSetting( SKeys::Setting, SSubKeys::SaveWindowPositions ).toBool() )
-        this->restoreGeometry( Settings::getSetting( SKeys::Positions, this->metaObject()->className() ).toByteArray() );
+        this->restoreGeometry( Settings::getSetting( SKeys::Positions, "ReMix" ).toByteArray() );
 
     //Setup Objects.
     Settings::setInstance( new Settings( this ) );
@@ -69,7 +69,7 @@ ReMix::ReMix(QWidget* parent) :
 ReMix::~ReMix()
 {
     if ( Settings::getSetting( SKeys::Setting, SSubKeys::SaveWindowPositions ).toBool() )
-        Settings::setSetting( this->saveGeometry(), SKeys::Positions, this->metaObject()->className() );
+        Settings::setSetting( this->saveGeometry(), SKeys::Positions, "ReMix" );
 
     if ( trayObject != nullptr )
         trayObject->deleteLater();