+++ src/thread/writethread.hpp
@@ -14,7 +14,7 @@ class WriteThread : public QThread
 {
     Q_OBJECT
 
-    QMap<LKeys, QString> logType;
+    static const QMap<LKeys, QString> logType;
 
     QString logDate{ "" };
     QFile punishmentLog;
@@ -29,17 +29,17 @@ class WriteThread : public QThread
     QFile pingLog;
 
     public:
-        WriteThread(const QMap<LKeys, QString>& types, QObject *parent = nullptr);
+        WriteThread(QObject *parent = nullptr);
         ~WriteThread() override;
 
         void run() override;
 
-        static WriteThread* getNewWriteThread(const QMap<LKeys, QString>& types, QObject* parent = nullptr);
+        static WriteThread* getNewWriteThread(QObject* parent = nullptr);
 
         void logToFile(const LKeys& type, const QString& text, const QString& timeStamp, const bool& newLine);
         bool isLogOpen(const LKeys& type);
         QFile& getLogFile(const LKeys& type);
-        void openLogFile(const LKeys& type);
+        bool openLogFile(const LKeys& type);
         void closeLogFile(QFile& log);
         void closeAllLogFiles();